touriquest-backend/services/ai-service/pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai-service"
version = "0.1.0"
description = "TouriQuest AI Assistant Service with Google Gemini and Voice Processing"
authors = [ "TouriQuest Team <dev@touriquest.com>",]
readme = "README.md"
package-mode = false

[tool.black]
line-length = 88
target-version = [ "py311",]
include = "\\.pyi?$"

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
celery = "^5.3.4"
websockets = "^12.0"
httpx = "^0.25.2"
google-generativeai = "^0.3.2"
openai-whisper = "^20231117"
gtts = "^2.5.1"
langdetect = "^1.0.9"
sentence-transformers = "^2.2.2"
numpy = "^1.24.3"
scipy = "^1.11.4"
scikit-learn = "^1.3.2"
soundfile = "^0.12.1"
librosa = "^0.10.1"
torch = "^2.1.1"
transformers = "^4.35.2"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
pillow = "^10.1.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
tenacity = "^8.2.3"
prometheus-client = "^0.19.0"
structlog = "^23.2.0"
typing-extensions = "^4.8.0"

[tool.pytest.ini_options]
testpaths = [ "tests",]
python_files = [ "test_*.py", "*_test.py",]
python_functions = [ "test_*",]
addopts = "--cov=app --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard",]
version = "^0.24.0"

[tool.poetry.dependencies.redis]
extras = [ "hiredis",]
version = "^5.0.1"

[tool.poetry.dependencies.python-jose]
extras = [ "cryptography",]
version = "^3.3.0"

[tool.poetry.dependencies.passlib]
extras = [ "bcrypt",]
version = "^1.7.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
isort = "^5.12.0"
pre-commit = "^3.5.0"
httpx = "^0.25.2"
